Move parallel planning into planning package
