+++ moveit_ros/planning/planning_pipeline_interfaces/include/moveit/planning_pipeline_interfaces/stopping_criterion_functions.hpp
@@ -1,4 +1,3 @@
-
 /*********************************************************************
  * Software License Agreement (BSD License)
  *
@@ -36,31 +35,17 @@
 /* Author: Sebastian Jahr
    Desc: TODO */
 
-#include <moveit/pipeline_planning_interface/plan_responses_container.hpp>
+#pragma once
+
+#include <moveit/planning_pipeline_interfaces/planning_pipeline_interfaces.hpp>
 
 namespace moveit
 {
-namespace planning_interface
+namespace planning_pipeline_interfaces
 {
-PlanResponsesContainer::PlanResponsesContainer(const size_t expected_size)
-{
-  solutions_.reserve(expected_size);
-}
-
-/** \brief Thread safe method to add PlanResponsesContainer to this data structure TODO(sjahr): Refactor this method to
- * an insert method similar to https://github.com/ompl/ompl/blob/main/src/ompl/base/src/ProblemDefinition.cpp#L54-L161.
- * This way, it is possible to create a sorted container e.g. according to a user specified criteria
- */
-void PlanResponsesContainer::pushBack(const ::planning_interface::MotionPlanResponse& plan_solution)
-{
-  std::lock_guard<std::mutex> lock_guard(solutions_mutex_);
-  solutions_.push_back(plan_solution);
-}
-
-/** \brief Get solutions */
-const std::vector<::planning_interface::MotionPlanResponse>& PlanResponsesContainer::getSolutions() const
-{
-  return solutions_;
-}
-}  // namespace planning_interface
+/** \brief A callback function that can be used as a parallel planning stop criterion.
+ *          It stops parallel planning as soon as any planner finds a solution. */
+bool stopAtFirstSolution(const PlanResponsesContainer& plan_responses_container,
+                         const std::vector<::planning_interface::MotionPlanRequest>& plan_requests);
+}  // namespace planning_pipeline_interfaces
 }  // namespace moveit
