+++ moveit_ros/planning/planning_pipeline_interfaces/src/plan_responses_container.cpp
@@ -1,3 +1,4 @@
+
 /*********************************************************************
  * Software License Agreement (BSD License)
  *
@@ -35,34 +36,31 @@
 /* Author: Sebastian Jahr
    Desc: TODO */
 
-#pragma once
-
-#include <moveit/planning_interface/planning_response.h>
-#include <moveit/planning_interface/planning_request.h>
+#include <moveit/planning_pipeline_interfaces/plan_responses_container.hpp>
 
 namespace moveit
 {
-namespace planning_interface
+namespace planning_pipeline_interfaces
 {
-MOVEIT_CLASS_FORWARD(PlanResponsesContainer);  // Defines PlanningComponentPtr, ConstPtr, WeakPtr... etc
-/** \brief A container to thread-safely store multiple MotionPlanResponses for later usage */
-class PlanResponsesContainer
+PlanResponsesContainer::PlanResponsesContainer(const size_t expected_size)
 {
-public:
-  PlanResponsesContainer(const size_t expected_size = 0);
+  solutions_.reserve(expected_size);
+}
 
-  /** \brief Thread safe method to add PlanResponsesContainer to this data structure TODO(sjahr): Refactor this method to an
-   * insert method similar to https://github.com/ompl/ompl/blob/main/src/ompl/base/src/ProblemDefinition.cpp#L54-L161.
-   * This way, it is possible to create a sorted container e.g. according to a user specified criteria
-   */
-  void pushBack(const ::planning_interface::MotionPlanResponse& plan_solution);
+/** \brief Thread safe method to add PlanResponsesContainer to this data structure TODO(sjahr): Refactor this method to
+ * an insert method similar to https://github.com/ompl/ompl/blob/main/src/ompl/base/src/ProblemDefinition.cpp#L54-L161.
+ * This way, it is possible to create a sorted container e.g. according to a user specified criteria
+ */
+void PlanResponsesContainer::pushBack(const ::planning_interface::MotionPlanResponse& plan_solution)
+{
+  std::lock_guard<std::mutex> lock_guard(solutions_mutex_);
+  solutions_.push_back(plan_solution);
+}
 
-  /** \brief Get solutions */
-  const std::vector<::planning_interface::MotionPlanResponse>& getSolutions() const;
-
-private:
-  std::vector<::planning_interface::MotionPlanResponse> solutions_;
-  std::mutex solutions_mutex_;
-};
-}  // namespace planning_interface
+/** \brief Get solutions */
+const std::vector<::planning_interface::MotionPlanResponse>& PlanResponsesContainer::getSolutions() const
+{
+  return solutions_;
+}
+}  // namespace planning_pipeline_interfaces
 }  // namespace moveit
