+++ moveit_ros/planning_interface/CMakeLists.txt
@@ -31,7 +31,6 @@ include(ConfigExtras.cmake)
 
 set(THIS_PACKAGE_INCLUDE_DIRS
   common_planning_interface_objects/include
-  pipeline_planning_interface/include
   planning_scene_interface/include
 #  py_bindings_tools/include
   move_group_interface/include
@@ -40,7 +39,6 @@ set(THIS_PACKAGE_INCLUDE_DIRS
 set(THIS_PACKAGE_LIBRARIES
   moveit_common_planning_interface_objects
   moveit_move_group_interface
-  moveit_pipeline_planning_interface
   moveit_planning_scene_interface
 #  moveit_py_bindings_tools
 #  moveit_py_bindings_tools_python
@@ -67,7 +65,6 @@ include_directories(${THIS_PACKAGE_INCLUDE_DIRS})
 
 add_subdirectory(common_planning_interface_objects)
 add_subdirectory(move_group_interface)
-add_subdirectory(pipeline_planning_interface)
 add_subdirectory(planning_scene_interface)
 # add_subdirectory(py_bindings_tools)
 # add_subdirectory(robot_interface)
